feature: Default setting for annotation
