+++ package.json
@@ -3,51 +3,49 @@
   "version": "10.1.0",
   "private": true,
   "scripts": {
-  "dev": "next dev -H 0.0.0.0 -p 8180",
-  "build": "next build",
-  "start": "next start -H 0.0.0.0 -p 8180",
-  "format": "prettier --write \"{app,layout,types}/**/*.{js,ts,tsx,d.ts}\"",
-  "lint": "next lint",
-  "android-dev": "adb reverse tcp:80 tcp:8180 && react-native run-android"
+    "dev": "next dev -H 0.0.0.0 -p 8180",
+    "build": "next build",
+    "start": "next start -H 0.0.0.0 -p 8180",
+    "format": "prettier --write \"{app,layout,types}/**/*.{js,ts,tsx,d.ts}\"",
+    "lint": "next lint",
+    "android-dev": "adb reverse tcp:80 tcp:8180 && react-native run-android"
   },
   "dependencies": {
-  "@reduxjs/toolkit": "^2.2.5",
-  "@types/node": "20.3.1",
-  "@types/react": "^18.3.3",
-  "@types/react-dom": "18.2.5",
-  "@types/three": "^0.165.0",
-  "axios": "^1.7.2",
-  "chart.js": "^3.9.1",
-  "chartjs-adapter-luxon": "^1.3.1",
-  "chartjs-plugin-streaming": "^2.0.0",
-  "cors": "^2.8.5",
-  "formik": "^2.4.6",
-  "fs": "^0.0.1-security",
-  "install": "^0.13.0",
-  "luxon": "^3.4.4",
-  "next": "^14.2.4",
-  "next-redux-wrapper": "^8.1.0",
-  "nipplejs": "^0.10.2",
-  "primeflex": "^3.3.1",
-  "primeicons": "^6.0.1",
-  "primereact": "^10.6.5",
-  "react": "18.2.0",
-  "react-chartjs-2": "^4.3.1",
-  "react-dom": "18.2.0",
-  "react-redux": "^9.1.2",
-  "socket.io": "^4.7.5",
-  "socket.io-client": "^4.7.5",
-  "three": "^0.165.0",
-  "typescript": "5.1.3",
-  "yup": "^1.4.0",
-  "zustand": "^4.5.2"
+    "@reduxjs/toolkit": "^2.2.5",
+    "@types/node": "20.3.1",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "18.2.5",
+    "@types/three": "^0.165.0",
+    "axios": "^1.7.2",
+    "chart.js": "^3.9.1",
+    "chartjs-adapter-luxon": "^1.3.1",
+    "chartjs-plugin-streaming": "^2.0.0",
+    "cors": "^2.8.5",
+    "formik": "^2.4.6",
+    "fs": "^0.0.1-security",
+    "install": "^0.13.0",
+    "luxon": "^3.4.4",
+    "next": "^14.2.4",
+    "next-redux-wrapper": "^8.1.0",
+    "nipplejs": "^0.10.2",
+    "primeflex": "^3.3.1",
+    "primeicons": "^6.0.1",
+    "primereact": "^10.7.0",
+    "react": "18.2.0",
+    "react-chartjs-2": "^4.3.1",
+    "react-dom": "18.2.0",
+    "react-redux": "^9.1.2",
+    "socket.io": "^4.7.5",
+    "socket.io-client": "^4.7.5",
+    "three": "^0.165.0",
+    "typescript": "5.1.3",
+    "yup": "^1.4.0",
+    "zustand": "^4.5.2"
   },
   "devDependencies": {
-  "eslint": "8.43.0",
-  "eslint-config-next": "13.4.6",
-  "prettier": "2.8.8",
-  "sass": "^1.63.4"
+    "eslint": "8.43.0",
+    "eslint-config-next": "13.4.6",
+    "prettier": "2.8.8",
+    "sass": "^1.63.4"
   }
-  }
-  
-  
+}
