+++ styles/components/_map.scss
@@ -16,37 +16,40 @@
   flex-wrap: wrap;
   height: 70vh;
 }
-#loc-container {
+
+#property-container {
   position: relative;
-  right: 15px;
-  top: 15px;
-  z-index: 999;
-  #loc-panel {
-    min-width: 15vw;
+  right: 1rem;
+  top: 1rem;
+  .p-panel {
     position: absolute;
+    max-width: 20vw;
+    min-width: 15vw;
     right: 0;
-    opacity: 0.66;
-    #switch-container {
-      display: flex;
-      flex-wrap: wrap;
-      flex-direction: row;
-      align-items: center;
-      margin-bottom: 1rem;
-      .p-selectbutton {
-        margin-left: 0.5rem;
-      }
-    }
   }
-  .p-card-content {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
+}
+
+#switch-container {
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: row;
+  align-items: center;
+  margin-bottom: 1rem;
+  .p-selectbutton {
+    margin-left: 0.5rem;
   }
+}
+.p-card-content {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
 }
 
 .p-speeddial-item {
   .pi-map,
-  .pi-download {
+  .pi-download,
+  .pi-flag,
+  .pi-compass {
     color: #ffffff;
   }
 }
